ai-app-builder/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.0
httpx==0.25.2
stripe==7.8.0
jinja2==3.1.2
aiofiles==23.2.1
docker==6.1.3
gitpython==3.1.40
requests==2.31.0